+++ app/src/main/java/emi/diary_app/ShareActivity.java
@@ -1,0 +1,13 @@
+package emi.diary_app;
+
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class ShareActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_share);
+    }
+}
