+++ app/src/main/res/layout/diary_entry_voice.xml
@@ -21,7 +21,7 @@
         >
 
         <SeekBar
-            style="?android:attr/seekBarStyle"
+            style="@style/SeekBarAppTheme"
             android:layout_width="match_parent"
             android:textIsSelectable="false"
             android:layout_weight="1"
