--- app/src/main/res/menu/menu_click_on_entry.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    >
-
-    <item
-        android:id="@+id/showEntry"
-        android:title="Anzeigen" />
-    <item
-        android:id="@+id/deleteEntry"
-        android:title="Löschen" />
-    <item
-        android:id="@+id/shareEntry"
-        android:title="Share" />
-
-</menu>
